+++ packages/artifact-kernel-wasm-presentation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @opengeni/artifact-kernel-wasm-presentation
 
+## 0.1.1
+
 ## 0.1.0
 
 ### Minor Changes