+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opengeni/storage
 
+## 0.2.84
+
+### Patch Changes
+
+- Updated dependencies [2cd6dce]
+  - @opengeni/contracts@0.42.1
+  - @opengeni/config@0.12.10
+
 ## 0.2.83
 
 ### Patch Changes